move LocalNotificationHandler and related classes to core 2.0
