+++ code/android-core-library/src/phone/java/com/adobe/marketing/mobile/services/ui/AndroidUIService.java
@@ -21,6 +21,7 @@ import android.content.Intent;
 import android.net.Uri;
 import android.os.Build;
 
+import com.adobe.marketing.mobile.LocalNotificationHandler;
 import com.adobe.marketing.mobile.services.Log;
 import com.adobe.marketing.mobile.services.ServiceConstants;
 import com.adobe.marketing.mobile.services.ServiceProvider;
@@ -190,8 +191,7 @@ public class AndroidUIService implements UIService {
         }
 
         final Intent intent = new Intent(Intent.ACTION_VIEW);
-        // Todo: Fix after refactoring LocalNotificationHandler
-        //intent.setClass(appContext, LocalNotificationHandler.class);
+        intent.setClass(appContext, LocalNotificationHandler.class);
         intent.putExtra(NOTIFICATION_SENDER_CODE_KEY, NOTIFICATION_SENDER_CODE);
         intent.putExtra(NOTIFICATION_IDENTIFIER_KEY, notificationSetting.getIdentifier());
         intent.putExtra(NOTIFICATION_REQUEST_CODE_KEY, requestCode);
